[buildsys] don't copy but link source files into build directories
Not really large difference, but linking lots of files is more
efficient than copying.  Source files aren't changed during build.

This change depends on the previous (don't copy .dotfiles), since
one of possible .dotfiles to be found in (unclean) top source dir
is .config, which should not be linked.  But we don't include
.dotfiles into the list of files to be linked/copied anymore.