stop-checking-ancient-kernel-version.patch (#684611)
a rather trivial patch to disable compatibility code for kernels <2.2.18
in nfs mount.  This was the only place in debian busybox which used
get_linux_version_code() function which was buggy on kernels with
less than 3 components version number.  So instead of fixing that
function, we got rid of the last user of it.  (Closes: #684611)